+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Literate.jl changelog
 
+## Version `v2.8`
+
+* ![Feature][badge-feature] Execution of notebooks now capture output of `display(x)`
+  and `display(mime, x)` ([ceff7a3][ceff7a3]).
+
 ## Version `v2.7`
 
 * ![Feature][badge-feature] Multiline-style Julia comments (`#= ... =#`) can now be
@@ -87,6 +92,7 @@
 [e08ca0a]: https://github.com/fredrikekre/Literate.jl/commit/e08ca0a19bd5e61dac778ddf4aaf6cef37532e48
 [6d1aec9]: https://github.com/fredrikekre/Literate.jl/commit/6d1aec90b13c6ad888be0fdc77583e9c525b5dc1
 [dc409d0]: https://github.com/fredrikekre/Literate.jl/commit/dc409d0f43a6282bee4e28e8e12bb6309942e5d5
+[ceff7a3]: https://github.com/fredrikekre/Literate.jl/commit/ceff7a36be2a9152d853257bac97be00d915ba8e
 
 [badge-breaking]: https://img.shields.io/badge/BREAKING-red.svg
 [badge-deprecation]: https://img.shields.io/badge/deprecation-orange.svg